[WIP] Fix some bug on account_domain_mutes

+++ db/migrate/20230221054049_create_account_domain_mutes.rb
@@ -2,7 +2,7 @@ class CreateAccountDomainMutes < ActiveRecord::Migration[6.1]
   def change
     create_table :account_domain_mutes do |t|
       t.string :domain
-      t.integer :account_id
+      t.bigint :account_id
       t.boolean :hide_notifications
       t.boolean :hide_from_home
